# contracts/http/openapi.yaml
{
  "components": {
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Bad request"
      },
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Permission denied or CSRF rejected"
      },
      "InternalError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Unexpected Java baseline failure"
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Resource not found"
      },
      "TooManyRequests": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Rate limit exceeded"
      },
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Authentication required"
      }
    },
    "schemas": {
      "Error": {
        "additionalProperties": true,
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "csrfHeader": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "sessionCookie": {
        "in": "cookie",
        "name": "MP_SESSION",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "MusicParty frozen Java HTTP contract",
    "version": "go-rewrite-baseline-v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/command": {
      "post": {
        "operationId": "AdminController_handleAdminCommand",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Java baseline success response"
          },
          "204": {
            "description": "Java baseline success without a body"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "tags": [
          "AdminController"
        ],
        "x-java-source": "src/main/java/org/thornex/musicparty/controller/AdminController.java:66"
      }
    },
    "/api/admin/navidrome-access/grant": {
      "post": {
        "operationId": "AdminController_grantNavidromeAccess",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Java baseline success response"
          },
          "204": {
            "description": "Java baseline success without a body"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "tags": [
          "AdminController"
        ],
        "x-java-source": "src/main/java/org/thornex/musicparty/controller/AdminController.java:195"
      }
    },
    "/api/admin/navidrome-access/revoke": {
      "post": {
        "operationId": "AdminController_revokeNavidromeAccess",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Java baseline success response"
          },
          "204": {
            "description": "Java baseline success without a body"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "tags": [
          "AdminController"
        ],
        "x-java-source": "src/main/java/org/thornex/musicparty/controller/AdminController.java:203"
      }
    },
...
    "/api/admin/subsonic-source": {
      "post": {
        "operationId": "AdminController_upsertSubsonicSource",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Java baseline success response"
          },
          "204": {
            "description": "Java baseline success without a body"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "tags": [
          "AdminController"
        ],
        "x-java-source": "src/main/java/org/thornex/musicparty/controller/AdminController.java:171"
      }
    },
    "/api/admin/subsonic-source/order": {
      "post": {
        "operationId": "AdminController_reorderSubsonicSource",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Java baseline success response"
          },
          "204": {
            "description": "Java baseline success without a body"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "tags": [
          "AdminController"
        ],
        "x-java-source": "src/main/java/org/thornex/musicparty/controller/AdminController.java:240"
      }
    },
    "/api/admin/subsonic-source/remove": {
      "post": {
        "operationId": "AdminController_removeSubsonicSource",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Java baseline success response"
          },
          "204": {
            "description": "Java baseline success without a body"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "tags": [
          "AdminController"
        ],
        "x-java-source": "src/main/java/org/thornex/musicparty/controller/AdminController.java:260"
      }
    },
    "/api/admin/subsonic-source/test": {
      "post": {
        "operationId": "AdminController_testSubsonicSource",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Java baseline success response"
          },
          "204": {
            "description": "Java baseline success without a body"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "tags": [
          "AdminController"
        ],
        "x-java-source": "src/main/java/org/thornex/musicparty/controller/AdminController.java:222"
      }
    },
    "/api/admin/subsonic-sources": {
      "get": {
        "operationId": "AdminController_listSubsonicSources",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Java baseline success response"
          },
          "204": {
            "description": "Java baseline success without a body"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "tags": [
          "AdminController"
        ],
        "x-java-source": "src/main/java/org/thornex/musicparty/controller/AdminController.java:211"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
}